Improve SARIF output
Previously, the script would take the output of csdiff, pipe it to
csgrep and then pipe *that* to csgrep again to generate the SARIF
output. Some information gets lost in that process.

Instead, always pass the csdiff output to csgrep directly.

More specifically, it seems that the info that can get lost is the
column range of the warnings (output generated with sarif-fmt):

    error: Double quote array expansions to avoid re-splitting elements.
        ┌─ task/buildah/0.1/buildah.yaml:269:9
        │
    269 │         ${LABELS[@]} \
        │         ^

    error: Double quote array expansions to avoid re-splitting elements.
       ┌─ task/apply-tags/0.1/apply-tags.yaml:34:22
       │
    34 │           for tag in $@; do
       │                      ^

versus:

    error: Double quote array expansions to avoid re-splitting elements.
        ┌─ task/buildah/0.1/buildah.yaml:269:9
        │
    269 │         ${LABELS[@]} \
        │         ^^^^^^^^^^^^

    error: Double quote array expansions to avoid re-splitting elements.
       ┌─ task/apply-tags/0.1/apply-tags.yaml:34:22
       │
    34 │           for tag in $@; do
       │                      ^^

Signed-off-by: Adam Cmiel <acmiel@redhat.com>