checkton.py: rework offset handling
Previously, checkton would generate script files like this:

    #!/bin/bash

        set -euo pipefail
        echo $HELLO

I.e. most of the script would preserve the line and column offset as in
the original YAML file, but the shebang would always be on the first
line. That means errors in the shebang itself would always get reported
as if they were on the first line of the YAML file instead of getting
reported at the proper place.

Rework the offset handling. The script will generate files like this:

    #!/bin/bash
    set -euo pipefail
    echo $HELLO

I.e. just take the content of the script block and make a file out of
it. Checkton remembers the original line and column offset, and updates
the line and column numbers in the ShellCheck JSON accordingly.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>