Add workflow for building image
On every merge to main, build the image, tag it with the commit SHA and
push to ghcr.io

When a commit receives an 'image-vx.y.z' tag, build the image, tag it
with the SHA and version. Then, automatically open a PR to bump the
image version in action.yaml.

Also, always generate an attested build provenance and SBOM for the
image.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>